# docs/openapi-solicitudes.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"tags":[{"name":"Categorias","description":"Catalogo de categorias activas"},{"name":"Solicitudes","description":"Registro, consulta y transiciones de solicitudes"}],"paths":{"/api/v1/solicitudes":{"get":{"tags":["Solicitudes"],"summary":"Bandeja paginada. El SOLICITANTE solo ve las propias","operationId":"listar","parameters":[{"name":"estado","in":"query","required":false,"schema":{"type":"string"}},{"name":"categoriaId","in":"query","required":false,"schema":{"type":"string","format":"uuid"}},{"name":"prioridad","in":"query","required":false,"schema":{"type":"string"}},{"name":"desde","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"hasta","in":"query","required":false,"schema":{"type":"string","format":"date-time"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PaginaResponseSolicitudResponse"}}}}}},"post":{"tags":["Solicitudes"],"summary":"Registra una solicitud nueva (escenario A1)","operationId":"crear","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CrearSolicitudRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolicitudResponse"}}}}}}},"/api/v1/solicitudes/{id}/transiciones":{"post":{"tags":["Solicitudes"],"summary":"Resolver, devolver o cerrar (escenario A4 si la transicion no aplica)","operationId":"transicionar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TransicionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolicitudResponse"}}}}}}},"/api/v1/solicitudes/{id}/observaciones":{"post":{"tags":["Solicitudes"],"summary":"Agrega una observacion al expediente","operationId":"observar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ObservacionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolicitudResponse"}}}}}}},"/api/v1/solicitudes/{id}/asignaciones":{"post":{"tags":["Solicitudes"],"summary":"Un analista toma la solicitud (escenario A2)","operationId":"asignar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolicitudResponse"}}}}}}},"/api/v1/solicitudes/{id}":{"get":{"tags":["Solicitudes"],"summary":"Detalle con historial y observaciones","operationId":"detalle","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SolicitudResponse"}}}}}}},"/api/v1/categorias":{"get":{"tags":["Categorias"],"summary":"Lista las categorias activas para el formulario de creacion","operationId":"listar_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CategoriaResponse"}}}}}}}}},"components":{"schemas":{"CrearSolicitudRequest":{"type":"object","properties":{"asunto":{"type":"string","maxLength":200,"minLength":0},"descripcion":{"type":"string","maxLength":2000,"minLength":0},"categoriaId":{"type":"string","format":"uuid"},"prioridad":{"type":"string","pattern":"BAJA|MEDIA|ALTA"}},"required":["categoriaId","prioridad"]},"CambioEstadoResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"estadoOrigen":{"type":"string"},"estadoDestino":{"type":"string"},"actorId":{"type":"string"},"actorRol":{"type":"string"},"motivo":{"type":"string"},"ocurridoEn":{"type":"string","format":"date-time"}}},"ObservacionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"texto":{"type":"string"},"actorId":{"type":"string"},"actorRol":{"type":"string"},"ocurridoEn":{"type":"string","format":"date-time"}}},"SolicitudResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"codigo":{"type":"string"},"asunto":{"type":"string"},"descripcion":{"type":"string"},"categoriaId":{"type":"string","format":"uuid"},"prioridad":{"type":"string"},"estado":{"type":"string"},"solicitanteId":{"type":"string"},"analistaId":{"type":"string"},"creadaEn":{"type":"string","format":"date-time"},"actualizadaEn":{"type":"string","format":"date-time"},"historial":{"type":"array","items":{"$ref":"#/components/schemas/CambioEstadoResponse"}},"observaciones":{"type":"array","items":{"$ref":"#/components/schemas/ObservacionResponse"}}}},"TransicionRequest":{"type":"object","properties":{"accion":{"type":"string","pattern":"RESOLVER|DEVOLVER|CERRAR"},"motivo":{"type":"string","maxLength":500,"minLength":0}},"required":["accion"]},"ObservacionRequest":{"type":"object","properties":{"texto":{"type":"string","maxLength":1000,"minLength":0}}},"PaginaResponseSolicitudResponse":{"type":"object","properties":{"content":{"type":"array","items":{"$ref":"#/components/schemas/SolicitudResponse"}},"page":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"totalPages":{"type":"integer","format":"int32"}}},"CategoriaResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"codigo":{"type":"string"},"nombre":{"type":"string"}}}}}}